refactor(tui): extract diff rendering into state/diff.rs
First step of decomposing the 5.3k-line tui/state.rs. Move the
self-contained diff model and parsers — Line, DiffLine, unified_to_lines,
diff_lines, parse_rendered_diff — plus their four unit tests into a new
state/diff.rs submodule. state.rs becomes state/mod.rs.

Per the project's no-re-export rule, diff is a `pub mod` reached by
qualified path (state::diff::Line): ui.rs imports the diff items from
state::diff, and mod.rs pulls them into local scope with a private `use`
(no `pub use` facade). No behavior change.
